do a ton of refactoring
