+++ src/components/Inputs.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+import InputBox from "./InputBox";
+
+const Inputs = ({ handleChange, ...rest }) => {
+  const makeInputs = Object.entries(rest).map(input => (
+    <InputBox
+      key={input[0]}
+      className="center-block"
+      value={input[1]}
+      name={input[0]}
+      handleChange={handleChange}
+    >
+      {input[0]}
+    </InputBox>
+  ));
+  return <React.Fragment>{makeInputs}</React.Fragment>;
+};
+
+export default Inputs;
