+++ src/components/SpaceBkg.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { Rect } from "react-native-svg";
+
+const SpaceBkg = ({ x, y, galaxySize, stroke = "black", fill = "black" }) => {
+  return (
+    <Rect
+      x={x}
+      y={y}
+      width={galaxySize}
+      height={galaxySize}
+      stroke={stroke}
+      fill="black"
+    />
+  );
+};
+
+export default SpaceBkg;
