+++ src/consts/consts.js
@@ -1,0 +1,49 @@
+const blankStar = { x: 0, y: 0, radius: -1, fill: "black", name: "" };
+
+const fillArray = [
+  "aqua",
+  "coral",
+  "red",
+  "blue",
+  "green",
+  "yellow",
+  "gold",
+  "khaki",
+  "maroon",
+  "orangered",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white",
+  "white"
+];
+
+export { fillArray, blankStar };
