+++ src/util/util.js
@@ -47,7 +47,18 @@ function randIndex(array = []) {
   return array[Math.floor(Math.random() * array.length)];
 }
 
+function capitalize(s) {
+  return s[0].toUpperCase() + s.slice(1);
+}
+
 module.exports = {
+  randInt: randInt,
   randNumber: randNumber,
-  randIndex: randIndex
+  randNumberBias: randNumberBias,
+  randNumberBiasMin: randNumberBiasMin,
+  randNumberBiasMax: randNumberBiasMax,
+  randNumberBetaDist: randNumberBetaDist,
+  randBeta: randBeta,
+  randIndex: randIndex,
+  capitalize: capitalize
 };
